Definition of prepro,solver, and postpro for the elastostatic problem
